HEHEHE I'M HUMAN!

+++ Application/Source/Core/Timestep.h
@@ -3,19 +3,19 @@
 class Timestep
 {
 public:
-	Timestep(double time = 0.0f) : time_(time) {}
+	Timestep(float time = 0.0f) : time_(time) {}
 
-	inline operator double() const { return time_; }
+	inline operator float() const { return time_; }
 
-	inline double GetSeconds() const 
+	inline float GetSeconds() const 
 	{
 		return time_;
 	}
 
-	inline double GetMilliseconds() const
+	inline float GetMilliseconds() const
 	{
 		return time_ * 1000.f;
 	}
 private:
-	double time_;
+	float time_;
 };
